Move CelebrateLaunchModal to common/celebrate-launch/ subfolder
