+++ projects/packages/jetpack-mu-wpcom/src/common/celebrate-launch/celebrate-launch-modal.js
@@ -4,8 +4,8 @@ import { useCopyToClipboard } from '@wordpress/compose';
 import { useState, useEffect, createInterpolateElement } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import { Icon, copy } from '@wordpress/icons';
+import { wpcomTrackEvent } from '../tracks';
 import ConfettiAnimation from './confetti-animation';
-import { wpcomTrackEvent } from './tracks';
 
 import './celebrate-launch-modal.scss';
