+++ projects/packages/jetpack-mu-wpcom/src/common/celebrate-launch/celebrate-launch-modal.scss
