+++ projects/packages/jetpack-mu-wpcom/src/common/celebrate-launch/confetti-animation.ts
