+++ projects/packages/jetpack-mu-wpcom/src/features/launch-button/launch-button.js
@@ -3,7 +3,7 @@ import { getSiteData } from '@automattic/jetpack-script-data';
 import { __ } from '@wordpress/i18n';
 import { addQueryArgs } from '@wordpress/url';
 import { useState } from 'react';
-import CelebrateLaunchModal from '../../common/celebrate-launch-modal';
+import CelebrateLaunchModal from '../../common/celebrate-launch/celebrate-launch-modal';
 import { useLaunchSiteMutation } from '../../common/hooks';
 import { wpcomTrackEvent } from '../../common/tracks';
